# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is plantapp",
        "title": "Leaflove Mobil API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://leaflove.com.tr",
    "paths": {
...
        "/mobil/add-favorite": {
            "post": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Add Favorite, you need send one parameter in the parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorite"
                ],
                "summary": "Add Favorite",
                "parameters": [
                    {
                        "description": "Add Favorite",
                        "name": "fav",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Fav"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/delete-favorite": {
            "post": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Delete Favorite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorite"
                ],
                "summary": "Delete Favorite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plant Name",
                        "name": "plant_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/favorites": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get Favorites by user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorite"
                ],
                "summary": "Get Favorites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.MyFav"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/general-plants": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get General Plants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_plants"
                ],
                "summary": "Get General Plants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GeneralPlants"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/plant-detail": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get Plant DEtail by plant_user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plant"
                ],
                "summary": "Get PLant Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plant user id",
                        "name": "plant_user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.PlantResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/plant-upload": {
            "post": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Upload Plant Image by base64 parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plant"
                ],
                "summary": "Upload Plant Image",
                "parameters": [
                    {
                        "description": "Upload Plant Image",
                        "name": "imageData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ImageData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Plant"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mobil/plants": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get User Plant by user_id for my plants section",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plant"
                ],
                "summary": "Get User Plants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.MyPlant"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "definitions": {
...
        "controllers.Fav": {
            "type": "object",
            "properties": {
                "general_plants_id": {
                    "type": "integer"
                },
                "rose_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.ImageData": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                }
            }
        },
...
        "controllers.MyFav": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "plantId": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.PlantResult": {
            "type": "object",
            "properties": {
                "climate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "health": {
                    "type": "string"
                },
                "imageurl": {
                    "type": "string"
                },
                "soil": {
                    "type": "string"
                },
                "tfvname": {
                    "type": "string"
                },
                "uploadImage": {
                    "type": "string"
                },
                "uses": {
                    "type": "string"
                }
            }
        },
...
        "models.GeneralPlants": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Plant": {
            "type": "object",
            "properties": {
                "climate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "health": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "soil": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uses": {
                    "type": "string"
                }
            }
        },
...
    }
}